Attempt testing service
Currently not passing (or running, for that matter). Please check comments, and see if there is something to be fixed immediately.

Would most likely be easier to work with once we have some more logic to work with. Main problem now is mocking the execution of a query.

The query class is initalized, and setup with mocking data, however, I find it hard to mock the execution of the actual query, before comparing the output.